fix: suspense 삭제
